requirements.txt
streamlit
streamlit_chat
openai
python-dotenv
langchain
pinecone-client
tiktoken